dnn_from_scratch::report
Struct ReportData

pub struct ReportData { /* private fields */ }

Struct to store and manage training and testing data for each epoch during model training.

Implementations

impl ReportData

pub fn new(n_epochs: u32, error_metric: &str) -> ReportData

Creates a new ReportData object to store training and testing history.

Arguments

n_epochs: The total number of epochs for training.

error_metric: The type of error metric to track, either “accuracy” or “error”.

Returns

Returns a new ReportData instance.

pub fn add( &mut self, train_loss: f64, train_error: f64, test_loss: f64, test_error: f64, )

Adds a new set of metrics (train loss, train error, test loss, test error) for the current epoch.

Arguments

train_loss: The loss value for the training data in the current epoch.

train_error: The error value for the training data in the current epoch.

test_loss: The loss value for the test data in the current epoch.

test_error: The error value for the test data in the current epoch.

pub fn get_errors(&self) -> (Vec<f64>, Vec<f64>)

Retrieves the errors (training and testing errors) from the report.

Returns

Returns a tuple containing two vectors:

Training errors as Vec<f64>.

Testing errors as Vec<f64>.

pub fn get_losses(&self) -> (Vec<f64>, Vec<f64>)

Retrieves the losses (training and testing losses) from the report.

Returns

Returns a tuple containing two vectors:

Training losses as Vec<f64>.

Testing losses as Vec<f64>.

pub fn is_empty(&self) -> bool

Checks if the ReportData is empty (i.e., no data has been added yet).

Returns

Returns true if the report is empty, false otherwise.

pub fn print_report(&self, epoch: u32)

Prints a formatted report for the current epoch, including training and testing loss/error/accuracy.

The report displays metrics in a table format, where the training and testing losses, as well as errors/accuracies, are printed for the current epoch.

Arguments

epoch: The current epoch number to display.

pub fn save_training_history(&self, output_path: &str)

Saves the entire training history (losses and errors) to a specified file.

The file is created if it doesn’t exist, and new entries are appended.

Arguments

output_path: The path to the output file where the training history will be saved.

pub fn save_report(&self, output_file: &str)

Saves the training history and report to a file within a specific directory (./output).

Arguments

output_file: The name of the output file where the training report will be saved.