dnn_from_scratch::utils
Struct Classification

pub struct Classification {}

Represents the Classification struct, which provides methods related to classification tasks.

Implementations

impl Classification

pub fn argmax(arr: &Array2<f64>, axis: usize) -> Array1<usize>

Perform argmax on a 2D array along a given axis (0 for rows, 1 for columns).

This method returns the indices of the maximum values along the specified axis. It is commonly used in classification problems to determine the predicted class.

Arguments

arr: A 2D array where each row or column represents a set of values.

axis: The axis along which the operation is performed:

0: Perform argmax along rows.

1: Perform argmax along columns.

Returns

Returns a 1D array containing the indices of the maximum values for each row or column, depending on the axis.

pub fn compute_accuracy<T>(output: &ArrayD<T>, target: &ArrayD<T>) -> f64
where T: PartialEq,

Compute the accuracy of a classification model.

This method compares the output predictions with the target labels and computes the proportion of matching elements.

Arguments

output: A multi-dimensional array representing the predicted values.

target: A multi-dimensional array representing the true target values.

Returns

Returns a f64 value representing the accuracy of the predictions (range 0.0 to 1.0).